[urlgrabber] Fix my poor's man concurrency.
The class attribute Job.prototype.running was incremented in
the page.onLoadStarted event which is loooong after the Job
has been created.  Meanwhile, the loop creating the Job objects
for each URL keeps running and a lot of objects are created before
the "running" class attribute reaches the requested concurrency
level.

Move the increment of this variable in the the Job constructor
which is more sensible anyway.  (I think I put it initially in
page.onLoadStarted() to know how many webpage objects were
actually working.)